+++ src/test/java/io/fabric8/maven/docker/util/EnvUtilTest.java
@@ -175,14 +175,6 @@ public class EnvUtilTest {
 
     }
 
-    @Test
-    public void isValidWindowsFileName() {
-
-    	assertFalse(EnvUtil.isValidWindowsFileName("/Dockerfile"));
-    	assertTrue(EnvUtil.isValidWindowsFileName("Dockerfile"));
-    	assertFalse(EnvUtil.isValidWindowsFileName("Dockerfile/"));
-    }
-
     private Properties getTestProperties(String ... vals) {
         Properties ret = new Properties();
         for (int i = 0; i < vals.length; i+=2) {